python/requirements.txt
Pillow
cairosvg
numpy
RPi.GPIO
spidev
